# docs/OPENAPI_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Healthcare Services Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/verification/queue": {
      "get": {
        "tags": [
          "Admin Doctor Verification Pipeline"
        ],
        "summary": "Get Verification Queue",
        "description": "Retrieves doctors in the verification queue.\nBy default returns 'submitted', 'under_review', and 'info_requested'.",
        "operationId": "get_verification_queue_api_v1_admin_verification_queue_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filter_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VerificationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VerificationQueueItemRead"
                  },
                  "title": "Response Get Verification Queue Api V1 Admin Verification Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/{doctor_id}": {
      "get": {
        "tags": [
          "Admin Doctor Verification Pipeline"
        ],
        "summary": "Get Doctor Verification Details",
        "operationId": "get_doctor_verification_details_api_v1_admin_verification__doctor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doctor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doctor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationQueueItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/{doctor_id}/transition": {
      "post": {
        "tags": [
          "Admin Doctor Verification Pipeline"
        ],
        "summary": "Transition Verification Status",
        "description": "Transitions doctor verification status through the 6-state pipeline.\nMandates reason for rejections and suspensions.\nWrites immutable audit log entry in the same transaction (RUL-04).",
        "operationId": "transition_verification_status_api_v1_admin_verification__doctor_id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doctor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doctor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationQueueItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DoctorDocumentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctor_id",
          "doc_type",
          "file_name",
          "s3_key",
          "uploaded_at"
        ],
        "title": "DoctorDocumentRead"
      },
...
      "DoctorRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "medical_reg_number": {
            "type": "string",
            "title": "Medical Reg Number"
          },
          "council_name": {
            "type": "string",
            "title": "Council Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "years_experience": {
            "type": "integer",
            "title": "Years Experience"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "listing_online": {
            "type": "boolean",
            "title": "Listing Online"
          },
          "video_enabled": {
            "type": "boolean",
            "title": "Video Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "full_name",
          "medical_reg_number",
          "council_name",
          "specialty",
          "years_experience",
          "bio",
          "verification_status",
          "listing_online",
          "video_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "DoctorRead"
      },
...
      "DocumentType": {
        "type": "string",
        "enum": [
          "medical_reg_cert",
          "degree_cert",
          "clinic_reg",
          "identity_proof"
        ],
        "title": "DocumentType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationQueueItemRead": {
        "properties": {
          "doctor": {
            "$ref": "#/components/schemas/DoctorRead"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DoctorDocumentRead"
            },
            "type": "array",
            "title": "Documents",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "doctor"
        ],
        "title": "VerificationQueueItemRead"
      },
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "under_review",
          "info_requested",
          "verified",
          "rejected",
          "suspended"
        ],
        "title": "VerificationStatus",
        "description": "PRD Section 2D: 6-state doctor verification pipeline"
      },
      "VerificationTransitionRequest": {
        "properties": {
          "new_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "reason_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Text",
            "description": "Mandatory when status is rejected or suspended"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          }
        },
        "type": "object",
        "required": [
          "new_status"
        ],
        "title": "VerificationTransitionRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}